requirements.txt
python-telegram-bot[rate-limiter]==22.3
selenium==4.21.0
webdriver-manager==4.0.2
python-dotenv==1.0.1
